requirements.txt
opencv-python
pillow
moondream
matplotlib
numpy
tqdm
gradio
